requirements.txt
pyrogram
python-dotenv
tinify
pymongo
